task_2_api/requirements.txt
pytest==8.4.1
requests>=2.31.0,<3.0.0
allure-pytest==2.13.5
ruff==0.12.5
